install.sh: fix grep error.

+++ install.sh
@@ -157,11 +157,12 @@ __install_darktable() {
 # :target: deadbeef - DeaDBeeF music player configuration
 __install_deadbeef() {
     # Preserve tabs and playlists.
-    if [[ -f "${HOME}/.config/deadbeef/config" ]]
+    if [[ -f "${HOME}/.config/deadbeef/config" ]] &&
+        grep -q "^playlist.tab" "${HOME}/.config/deadbeef/config"
     then
-        grep "^playlist.tab" "${HOME}/.config/deadbeef/config" >"${HOME}/.config/deadbeef/tabs"
-        install --copy "deadbeef/.config/deadbeef/config"
-        cat "${HOME}/.config/deadbeef/tabs" >>"${HOME}/.config/deadbeef/config"
+            grep "^playlist.tab" "${HOME}/.config/deadbeef/config" >"${HOME}/.config/deadbeef/tabs"
+            install --copy "deadbeef/.config/deadbeef/config"
+            cat "${HOME}/.config/deadbeef/tabs" >>"${HOME}/.config/deadbeef/config"
     else
         install --copy "deadbeef/.config/deadbeef"
     fi
